first encoded line accounted for
the add on of 127 at the start of the first line looks kinda suspicious. Made it so that the first line has a possible max length that is shorter than the max length of every line after. 
I believe that this will take away some suspicion when the max length of the first line (31) is in the range of the length of every other line (1 -51).