fixed checkFormatExercise
it was a combination of the regex pattern used to seek and the Matcher.find() method. I used (), saying that everything between these has to be present in the string,
and then i used the Matcher.matches() method in order to check if ENTIRE string is the sequence in the regex pattern.