added methods, did first trial of tests
as titles says. I checked if DataStorage and DataUnit can work with each other to read from a file with nested units, and then print the contents out. further extensive and critical testing needed in order for application in a showcase. Also, a lot of methods are to be added after a bit more testing.