update on project
have been having some problems with implementing a good way to add units that aren't in the first layer. Will try using the protected modifier to achive it somehow. Or just abandon the concept cuz user can just make the entire unit before adding it to the DataStorage, idk yet.